Fix exception in chart lens caused by missing contract type
The chart lens uses a in memory default contract for setting up the
initial chart configuration. This contract did not have a type set,
which caused a downstream error that made the lens crash.
A long term fix for this kind of problem is to add correct type
annotations to the UI, but I will leave a later PR.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>