Use single purpose endpoints for actions performed by anonymous users
Using single purpose endpoints for actions performed by anonymous users
allow us to run the actions using a privileged user session and remove
the ability for the guest user to do it directly. This allows us to
reduce the scope of access that the guest user has, and therefore
anonymous users that the guest user represents.

Change-type: major
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>