UI: Remove special handling of repository contracts
This change removes the special handling for repository contracts.
Previously we were experimenting with the idea that the repository would
behave like a chat room or flow in flowdock, that focused conversation
around a specific repository. The approach we want to try now is to
focus conversation around specific products, and use tags to create
views for specific topics.

Change-type: major
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>