UI: Always show the link tab for the next entity type in the loop
When viewing a contract, if that contract is part of the loop structure,
always show the link tab for the next type in the loop flow. E.g. if
looking at a pattern, the link tab for improvements will always be shown

Fixes https://github.com/product-os/jellyfish/issues/7853

Change-type: minor
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>