UI: Don't remove $$links form queries when using Chart lens
The original idea behind this is that you can make the query run a lot
fast by not expanding additional data. Unfortunately this breaks if the
$$links query is essential for the query to make sense (e.g. milestones
linked to an improvement). The best case is that we can strip out links
that are optional, but I will leave this for a future improvement.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>