UI: Split my user lens into seperate view and edit modes
This changes your own user lens to function in a much more similar way to the rest of the app and adds tabs for bookmarks and loops.
To edit your account you click the pencil icon in the top right, just like you would with any other contract.

Change-type: minor
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>