Use session identity instead of tokens for worker and kernel
This change brings in the [update to autumndb](https://github.com/product-os/autumndb/pull/1363) that removes direct session handling.
Instead of AutumnDB handling session tokens, it only deals with the identity/actor making requests. Authorization of users is now hoisted up to the route layer at the top of the call stack, where it can eventually be delegated to an external service, such as Ory Kratos.

Change-type: major
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>