Run UI service lint and unit tests in UI dockerfile
Currently these tests are run inside the SUT dockerfile, which means
that image needs to load all of the UI source code as well. With this
change, the UI unit and lint tests are run as part of the UI service build step.
This approach makes more sense and should reduce the size and build time
of the SUT container, which is currently the biggest bottleneck in the
stack's build step.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>