UI: Implement new NavBar component
As part of transitioning to the new iteration of the Jellyfish UI the
user profile/avatar, search and loop selector will now appear in
a persistent navigation element at the top of the page.

In a follow up PR, the "loop selector" dropdown will be reworked into
a navigation control.

Connects-to: https://github.com/product-os/jellyfish/issues/8788
Change-type: minor
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>