Fix inbox to handle sent tab query results
Update the EventBox component used in the new Inbox to properly
handle cases in which the query results in a message/whisper contract
and stop assuming result is always a thread. Also add prop to toggle
archive button - should only be able to archive things from the open
tab.

Change-type: patch
Signed-off-by: Josh Bowling <josh@monarci.com>