UI: Prefer linked target to hardcoded target when clicking a message
This makes the on-click behaviour much more responsive to context than
it used to be, by relying on the linked data if it's available.
For example, the message may have initially targeted and entity that no
longer exists, but is being queried through a more recent link. By using
the link data, it is much more likely that clicking the message will
result in the correct behaviour.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>