UI: Improve image rendering
This change keeps the generated blob source data in cache in the
browser, reducing server load and visually disruptive re-renders.
This cache is ephemeeral and flushed on a browser navigation.
The ideal solution is to use cookies to authenticate files instead of
fetching blob data and assembling it in the browser.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>